# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChanAnalyzer API",
    "description": "A股缠论研究工作台 v2",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EmptyJobCreate"
                  },
                  {
                    "$ref": "#/components/schemas/RefreshJobCreate"
                  },
                  {
                    "$ref": "#/components/schemas/ScanJobCreate"
                  },
                  {
                    "$ref": "#/components/schemas/BatchAnalysisJobCreate"
                  },
                  {
                    "$ref": "#/components/schemas/StockAnalysisJobCreate"
                  },
                  {
                    "$ref": "#/components/schemas/DataRefreshJobCreate"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "market_radar.refresh": "#/components/schemas/EmptyJobCreate",
                    "limit_up.refresh": "#/components/schemas/RefreshJobCreate",
                    "ipo.refresh": "#/components/schemas/RefreshJobCreate",
                    "scan.buy": "#/components/schemas/ScanJobCreate",
                    "scan.sell": "#/components/schemas/ScanJobCreate",
                    "screen.hot": "#/components/schemas/ScanJobCreate",
                    "screen.smart": "#/components/schemas/ScanJobCreate",
                    "limit_up.analyze": "#/components/schemas/BatchAnalysisJobCreate",
                    "ipo.analyze": "#/components/schemas/BatchAnalysisJobCreate",
                    "stock.analyze": "#/components/schemas/StockAnalysisJobCreate",
                    "data.refresh": "#/components/schemas/DataRefreshJobCreate"
                  }
                },
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Recent Jobs",
        "operationId": "recent_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobView"
                  },
                  "title": "Response Recent Jobs Api V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Archive Job",
        "operationId": "archive_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/items": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Items",
        "operationId": "job_items_api_v1_jobs__job_id__items_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobItemView"
                  },
                  "title": "Response Job Items Api V1 Jobs  Job Id  Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/results": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Results",
        "operationId": "job_results_api_v1_jobs__job_id__results_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanResultView"
                  },
                  "title": "Response Job Results Api V1 Jobs  Job Id  Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/stages": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Stages",
        "operationId": "job_stages_api_v1_jobs__job_id__stages_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStagesView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/rerun": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Rerun Job",
        "operationId": "rerun_job_api_v1_jobs__job_id__rerun_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Events",
        "operationId": "job_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAnalysisJobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "limit_up.analyze",
              "ipo.analyze"
            ],
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/BatchAnalysisPayload"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind",
          "payload"
        ],
        "title": "BatchAnalysisJobCreate"
      },
...
      "DataRefreshJobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "data.refresh",
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/DataRefreshPayload"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "DataRefreshJobCreate"
      },
...
      "EmptyJobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "market_radar.refresh",
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/EmptyPayload"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "EmptyJobCreate"
      },
      "EmptyPayload": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "EmptyPayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobAccepted"
      },
      "JobItemView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subject_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Key"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_key",
          "instrument_id",
          "status",
          "attempts",
          "result",
          "error",
          "started_at",
          "finished_at"
        ],
        "title": "JobItemView"
      },
...
      "JobStagesView": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobStageView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "items"
        ],
        "title": "JobStagesView"
      },
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "retry_of_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "progress",
          "total",
          "completed",
          "failed",
          "message",
          "error",
          "result",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobView"
      },
...
      "RefreshJobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "limit_up.refresh",
              "ipo.refresh"
            ],
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/RefreshPayload"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "RefreshJobCreate"
      },
...
      "ScanJobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "scan.buy",
              "scan.sell",
              "screen.hot",
              "screen.smart"
            ],
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/ScanPayload"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ScanJobCreate"
      },
...
      "ScanResultView": {
        "properties": {
          "instrument_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "scan_kind": {
            "type": "string",
            "title": "Scan Kind"
          },
          "signal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Type"
          },
          "signal_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Date"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "scan_kind",
          "signal_type",
          "signal_date",
          "rank",
          "score",
          "payload"
        ],
        "title": "ScanResultView"
      },
...
      "StockAnalysisJobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "stock.analyze",
            "title": "Kind"
          },
          "payload": {
            "$ref": "#/components/schemas/StockAnalysisPayload"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind",
          "payload"
        ],
        "title": "StockAnalysisJobCreate"
      },
      "StockAnalysisPayload": {
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "StockAnalysisPayload"
      },
...
    }
  }
}